refactor(actions/tasks): co-locate settings and demo in subfolders
